*** empty log message ***
git-svn-id: svn://localhost/trunk@864 7f4d9e04-a92a-ab41-bea9-970b690ef4a7

+++ org.python.pydev.debug/src/org/python/pydev/debug/model/PyDebugTarget.java
@@ -76,8 +76,10 @@ public class PyDebugTarget extends AbstractDebugTarget {
         }
         
 		threads = new IThread[0];
-		process.terminate();
-        process = null;
+		if(process != null){
+			process.terminate();
+			process = null;
+		}
 		fireEvent(new DebugEvent(this, DebugEvent.TERMINATE));
 	}
